perf: Split `MbfSmoother` compilation by measurement dimension
