+++ Core/include/Acts/TrackFitting/MbfSmoother.hpp
@@ -103,6 +103,11 @@ class MbfSmoother {
   void visitMeasurement(const AnyConstTrackStateProxy& ts,
                         BoundMatrix& bigLambdaHat,
                         BoundVector& smallLambdaHat) const;
+
+  template <std::size_t N>
+  void visitMeasurementImpl(const AnyMutableTrackStateProxy& ts,
+                            BoundMatrix& bigLambdaHat,
+                            BoundVector& smallLambdaHat) const;
 };
 
 }  // namespace Acts
